feat: implement e-commerce features and enhancements
- Add product detail pages with dynamic routing
- Implement promotional sections with category-based product displays
- Enhance payment integration with Razorpay
- Improve admin order management with status updates
- Add image handling and storage functionality
- Update product and category data models
- Implement responsive product grids and cards
- Add footer with dynamic category loading
- Improve authentication and cart context
- Enhance UI with animations and scroll behavior

+++ backend/src/main/java/com/vijaybrothers/store/dto/ProductDto.java
@@ -69,7 +69,7 @@ public class ProductDto {
         dto.setPrice(product.getPrice());
         dto.setCategoryId(product.getCategory() != null ? product.getCategory().getCategoryId() : null);
         dto.setImages(product.getImages().stream()
-                .map(img -> new ProductImageDto(img.getId().intValue(), img.getImageUrl(), img.isMain()))
+                .map(img -> new ProductImageDto(img.getId(), img.getImageUrl(), img.isMain()))
                 .collect(Collectors.toList()));
         dto.setStockQuantity(product.getStockQuantity());
         dto.setInStock(product.getInStock());
